add && remove instrument for user

+++ database/migrations/2019_06_20_115746_create_instruments_history_table.php
@@ -17,7 +17,7 @@ class CreateInstrumentsHistoryTable extends Migration
             $table->bigIncrements('id');
             $table->bigInteger('user_id')->unsigned();
             $table->bigInteger('instrument_id')->unsigned();
-            $table->bigInteger('history_type_id')->unsigned();
+            $table->integer('history_type_id')->unsigned();
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users');
